app/requirements.txt
pillow==9.0.0
celery
tfa-nightly
google-api-python-client
google-cloud-storage
